fix(relatorios): COUNT(*)::int nos DTOs de quantidade — prevenção da classe de bug das comissões
Postgres retorna COUNT(*) como int8 (bigint). Dapper falha ao materializar
bigint em C# int sem cast explícito. Três queries em RelatorioQueryRepository
usavam COUNT(*) sem ::int para colunas mapeadas em int Quantidade:
  - FaturamentoCategoriaDto (relatório de faturamento por categoria)
  - AgendamentosPorStatusDto (relatório de agendamentos por status)
  - AgendamentosPorDiaDto (relatório de agendamentos por dia)

Mesma classe do bug 500 das comissões corrigido em 82ace55.

+4 testes NUnit: 3 de contrato de DTO (int Quantidade) + 1 de regressão
SQL (verifica ausência de COUNT(*) sem ::int e presença das 3 ocorrências
corrigidas — falha em build se o padrão errado for reintroduzido).

+++ backend/src/Services/Imedto.Backend.Infrastructure/Database/Repositories/RelatorioQueryRepository.cs
@@ -62,7 +62,7 @@ public class RelatorioQueryRepository
                 tipo                    AS Tipo,
                 COALESCE(SUM(CASE WHEN status = 'Pago'     THEN valor ELSE 0 END), 0) AS TotalPago,
                 COALESCE(SUM(CASE WHEN status = 'Pendente' THEN valor ELSE 0 END), 0) AS TotalPendente,
-                COUNT(*)                AS Quantidade
+                COUNT(*)::int           AS Quantidade
             FROM lancamentos
             WHERE estabelecimento_id = @EstabelecimentoId
               AND status != 'Cancelado'
@@ -89,7 +89,7 @@ public class RelatorioQueryRepository
 
         // Batch único: 1 round-trip ao Postgres em vez de 2.
         const string sqlBatch = """
-            SELECT status AS Status, COUNT(*) AS Quantidade
+            SELECT status AS Status, COUNT(*)::int AS Quantidade
             FROM agendamentos
             WHERE estabelecimento_id = @EstabelecimentoId
               AND (@DataInicio::date IS NULL OR inicio_previsto::date >= @DataInicio::date)
@@ -97,7 +97,7 @@ public class RelatorioQueryRepository
             GROUP BY status
             ORDER BY Quantidade DESC;
 
-            SELECT inicio_previsto::date AS Data, COUNT(*) AS Quantidade
+            SELECT inicio_previsto::date AS Data, COUNT(*)::int AS Quantidade
             FROM agendamentos
             WHERE estabelecimento_id = @EstabelecimentoId
               AND (@DataInicio::date IS NULL OR inicio_previsto::date >= @DataInicio::date)
